# packages/sie_gateway/openapi.json
{
  "components": {
    "schemas": {
      "AllItemsFailedResponse": {
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/InferenceErrorDetail"
            },
            "type": "array"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "details"
        ],
        "type": "object"
      },
...
      "BundleConflictResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/BundleRoutingConflictDetail"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
      "BundleRoutingConflictDetail": {
        "properties": {
          "code": {
            "type": "string"
          },
          "compatible_bundles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "compatible_bundles"
        ],
        "type": "object"
      },
...
      "DocumentInput": {
        "properties": {
          "data": {
            "items": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "format": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ErrorDetailCore": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ExtractParams": {
        "properties": {
          "instruction": {
            "type": [
              "string",
              "null"
            ]
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "options": {},
          "output_schema": {}
        },
        "type": "object"
      },
      "ExtractRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemInput"
            },
            "minItems": 1,
            "type": "array"
          },
          "params": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExtractParams"
              }
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "ExtractResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExtractResult"
            },
            "type": "array"
          },
          "model": {
            "type": "string"
          }
        },
        "required": [
          "model",
          "items"
        ],
        "type": "object"
      },
      "ExtractResult": {
        "properties": {
          "classifications": {
            "items": {},
            "type": "array"
          },
          "data": {},
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/Relation"
            },
            "type": "array"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "GatewayErrorResponse": {
        "properties": {
          "configured_gpu_types": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/InferenceErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "error": {},
          "gpu": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "GatewayModelLoadFailedDetail": {
        "description": "SDK-style ``502`` body for ``MODEL_LOAD_FAILED`` (SDK short-circuit).",
        "properties": {
          "attempts": {
            "format": "int32",
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "error_class": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "permanent": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "attempts",
          "permanent"
        ],
        "type": "object"
      },
      "GatewayModelLoadFailedResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/GatewayModelLoadFailedDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "GpuNotConfiguredError": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/GpuNotConfiguredDetail"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "ImageInput": {
        "properties": {
          "data": {
            "items": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "format": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "InferenceInternalServerErrorResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/AllItemsFailedResponse"
          },
          {
            "$ref": "#/components/schemas/StandardApiError"
          }
        ]
      },
      "InferenceServiceUnavailableResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/GatewayErrorResponse"
          },
          {
            "$ref": "#/components/schemas/StandardApiError"
          },
          {
            "$ref": "#/components/schemas/GpuNotConfiguredError"
          }
        ]
      },
      "ItemInput": {
        "properties": {
          "document": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentInput"
              }
            ]
          },
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageInput"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "metadata": {},
          "text": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "ProvisioningResponse": {
        "properties": {
          "bundle": {
            "type": "string"
          },
          "estimated_wait_s": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "gpu": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "estimated_wait_s"
        ],
        "type": "object"
      },
...
      "StandardApiError": {
        "description": "FastAPI-style error envelope for most gateway JSON errors.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetailCore"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "SIE bearer token. Mutating pool/config/admin routes require the configured admin token; other protected routes accept a normal gateway token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Rust gateway API for SIE inference, pool coordination, and read-only runtime config.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "SIE Gateway",
    "version": "0.3.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/extract/{model}": {
      "post": {
        "description": "Mixed-success batches return 200 with only successful items; the response carries no per-item error envelope. For per-item error visibility, send single-item batches.",
        "operationId": "proxy_extract",
        "parameters": [
          {
            "description": "Model id. Runtime route is Axum catch-all `/v1/extract/{*model}`; clients using this OpenAPI path template should percent-encode slashes in model ids, for example `BAAI%2Fbge-m3`.",
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Preferred GPU or machine profile",
            "in": "header",
            "name": "X-SIE-MACHINE-PROFILE",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "Explicit pool routing override",
            "in": "header",
            "name": "X-SIE-Pool",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "Client SDK version for skew warnings",
            "in": "header",
            "name": "X-SIE-SDK-Version",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            },
            "application/msgpack": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            },
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            },
            "application/x-msgpack": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              },
              "application/msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              },
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              },
              "application/x-msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            },
            "description": "Extract response",
            "headers": {
              "X-Inference-Time": {
                "description": "Worker-reported inference time in milliseconds",
                "schema": {
                  "type": "string"
                }
              },
              "X-Payload-Fetch-Time": {
                "description": "Worker-reported offloaded payload fetch time in milliseconds, when available",
                "schema": {
                  "type": "string"
                }
              },
              "X-Postprocessing-Time": {
                "description": "Worker-reported postprocessing time in milliseconds, when available",
                "schema": {
                  "type": "string"
                }
              },
              "X-Queue-Publish-Time": {
                "description": "Milliseconds spent publishing work to the queue",
                "schema": {
                  "type": "string"
                }
              },
              "X-Queue-Time": {
                "description": "Worker-reported queue time in milliseconds",
                "schema": {
                  "type": "string"
                }
              },
              "X-Queue-Wait-Time": {
                "description": "Milliseconds spent waiting for worker results",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Request-Id": {
                "description": "Gateway request id for queue-backed inference",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Server-Version": {
                "description": "Gateway-compatible server version advertised by this gateway",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Version": {
                "description": "Gateway package version that handled the request",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Worker": {
                "description": "Logical queue worker tag that produced the response",
                "schema": {
                  "type": "string"
                }
              },
              "X-Tokenization-Time": {
                "description": "Worker-reported tokenization time in milliseconds, when available",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningResponse"
                }
              }
            },
            "description": "Worker provisioning in progress",
            "headers": {
              "Retry-After": {
                "description": "Suggested retry delay in seconds",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Server-Version": {
                "description": "Gateway-compatible server version advertised by this gateway",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Version": {
                "description": "Gateway package version that handled the request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Model not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleConflictResponse"
                }
              }
            },
            "description": "Bundle override conflicts with model routing"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Request body too large"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InferenceInternalServerErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/StandardApiError"
                    }
                  ]
                }
              }
            },
            "description": "All batch items failed or gateway internal error; gateway auth enabled but no tokens configured"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayModelLoadFailedResponse"
                }
              }
            },
            "description": "Terminal model load failure (MODEL_LOAD_FAILED)",
            "headers": {
              "X-SIE-Error-Code": {
                "description": "SDK-stable gateway or worker error code",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Error-Version": {
                "description": "Gateway package version associated with the error envelope",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Server-Version": {
                "description": "Gateway-compatible server version advertised by this gateway",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Version": {
                "description": "Gateway package version that handled the request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceServiceUnavailableResponse"
                }
              }
            },
            "description": "Queue unavailable, GPU not configured, model loading, or capacity exhausted",
            "headers": {
              "Retry-After": {
                "description": "Suggested retry delay in seconds, when retryable",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Error-Code": {
                "description": "SDK-stable gateway or worker error code",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Server-Version": {
                "description": "Gateway-compatible server version advertised by this gateway",
                "schema": {
                  "type": "string"
                }
              },
              "X-SIE-Version": {
                "description": "Gateway package version that handled the request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardApiError"
                }
              }
            },
            "description": "Result channel closed"
          }
        },
        "tags": [
          "inference"
        ],
        "x-sie-axum-catch-all": "/v1/extract/{*model}"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}